Enable BTF Output
This requires debuginfo=2 to be set - which is default for dev profile
and for the user to pass the `-C link-arg=--btf` to btf-linker.

LLVM will produce BTF, but also leaves `.debug_info` sections in the
object file. Since these sections aren't useful in the BTF context, we
use a rust implementation of `objcopy --strip-debug` to remove these
sections. This would also enable BTF to be supported on release builds,
should debuginfo also be set to 2.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>