di: Sanitize generic names (in correct way)
The previous way of stripping (instead of sanitizing) generic names was
not correct. When we define `MyStruct<T>`, then each implementation like
`MyStruct<u32>` or `MyStruct<u64>` is in fact a separate type. So to
differentiate them in BTF, we need to keep all of them with distinct
names.

However, Linux kernel doesn't accept BTF type names with the following
characters: `<`, `>`, `::`, `,`, ` `. So we need to sanitize them.