di: Improve type name sanitization
First of all, change the name of `strip_generics` function to
`sanitize_type_name`. The goal of it is more general - it makes sure
that Rust type names are sanitized to be compatible with C.

Then, instead of using `replace`, iterate over all characters once and
replace characters unsupported by C with their hex representations.