di: Sanitize generic names in functions (subprograms)
Previously we were sanitizing only struct types
(`DW_tag_structure_type`). Here it's done also for functions
(`DW_tag_subprogram`).

This change also fixes the `sanitize_type_name` function to accept the
`_` character. It's a valid character in C types and converting it to
`_5F_` is not great for readability.