Test on nightly only, use linker-flavor=bpf-linker
https://github.com/rust-lang/rust/commit/b0ce4164f09be1221d004c4b84e49bdcbf973194
added stricter validation to rustc, defeating the previous trick of
pretending to be wasm-ld because the wasm-ld flavor isn't compatible
with the host target. This means bpf can no longer build with
non-nightly rustc.

Change the test runner to use the bpf-linker linker-flavor and override
the target cpu to prevent rustc from passing x86-64, which is rejected
by bpf-linker.

Change the CI workflow to run tests only on nightly, but separate
nightly and nightly-from-source so that both paths are tested (the
former is what folks will run locally).