log: append rather than truncate
When passing `-Clink-arg=--log-file=path-to-file` to rustc through
cargo, it's likely that the linker is going to be used multiple times
concurrently.  This isn't ideal because the logs interleave; it would be
better to change the flag semantics to --log-dir so that each linker
instance can have its own log file, but that is left for another day.