Enable fatal errors by default
Ignore errors emitted on calls to builtins which are spurious. These
builtins end up being lowered to libcalls to hidden functions, which are
totally OK. See https://reviews.llvm.org/D155894 and
https://github.com/aya-rs/aya/pull/698 for some additional discussion.

There doesn't seem to be a silver bullet here; BPF is fundamentally
adversarial to these memory intrinsics. We can revisit this once we have
DI and can more easily locate the source location that results in their
inclusion.