di: Avoid null pointer dereference when accessing DIType name
Before this change, an anonymous C struct present in a shim (written
in C), linked by bpf-linker to a Rust program, would trigger a
null dereference. An example of a struct triggering it is:

```c
typedef struct {
  int counter;
} atomic_t;
```

To avoid such footguns, provide a helper function for getting the
DIType name safely.

Fixes: #130