di: Allow to make Rust types anonymous
Rust does not provide a possibility of defining anonymous types,
but we need it to produce BTF acceptable by the kernel for BTF
maps.

As a workaround, use our DI sanitizer to make type anonymous if
it contains an `_anon` field. To not affect the layout of the
struct, it should be defined as `PhantomData`, e.g.

```rust
pub struct Foo {
    _anon: PhantomData<u8>,
    something: u32,
}
```