di: Allow to make Rust types anonymous
Rust does not provide a possibility of defining anonymous types,
but we need it to produce BTF acceptable by the kernel for BTF
maps.

As a workaround, use our DI sanitizer to make type anonymous if
it contains a field of type `AyaBtfMapMarker`. To not affect the
layout of the struct, it should be defined as `PhantomData`:

```rust
pub struct AyaBtfMapMarker(PhantomData<()>);
```

Previously, we were "fixing" this issue by hard-coding `HashMap`
as a type for which we remove the name. Since there are many
map types in Aya, and we might introduce more in the future,
having tha marker type is a better solution.