di: Sanitize data carrying enums correctly
Data carrying enums are causing a crash of the BPF Assembly Printer
pass. As a workaround, we were always replacing the elements of that
enum metadata with an empty MDNode.

However, #142 broke this behavior by calling `replace_elements` on a
wrong composite type - it was calling it on the inner one (data
carrying field), instead of the outer one (enum). This change fixes
it.