chore: Update to LLVM 19
Latest Rust nightly switched to LLVM 19, so we need to update in order
to support rustc-llvm-proxy. That includes an update of rustc-llvm-proxy
to the version supporting LLVM 19.

LLVM 19 packages don't provide shared libraries anymore, so the usage
of `llvm-sys/force-dynamic` feature is dropped. Instead, we just let
llvm-sys decide what kind of libraries to use.

llvm-sys fails to build without libpolly being installed. Packages from
apt.llvm.org are being built all at once, with one cmake build with
superset of options, then different binaries and libraries are being
included in different packages.

That results in `llvm-config --libname --link-static` mentioning libpolly,
even if it's not installed. The output of that command is being used in
build.rs of llvm-sys, so building llvm-sys on such system is complaining about
lack of libpolly.

For now, we just install libpolly in CI, but hopefully that nightmare
goes away once we switch to binstalls and ditch the system LLVM option.