test: Add an xtask for building a permanent BPF standard library
Rust standard library can be built on demand with `-Zbuild-std`, but it
doesn't stay permanently in the current toolchain's sysroot.

We want to support not only toolchains provided by rustup, which don't
provide BPF targets, but also custom toolchains, which have them enabled
and provide a prebuilt standard library for them.

To imitate such custom toolchain without building it entirely from
scratch, provide the `cargo xtask build-std` command, which uses
rustc-build-sysroot to build the standard library for the given BPF
target (by default, `bpfel-unknown-none`), and keeps it in the current
toochain's sysroot.