Prevent LLVM from inserting ksyms symbols; add flag to allow (#294)
Recent versions of LLVM insert calls to __bpf_trap, which adds a .ksyms
section[0]. This is not currently supported by aya[1] and inevitably
causes complex programs to fail to load.

This PR adds the command line switch which disables this behavior, but
is only supported by very recent versions of LLVM. I'm not entirely
sure how to check that the LLVM version used supports this, so I've
marked this as draft.

Additionally, this behaviour is probably desired once .ksyms is
supported in aya, so this should be reverted when this happens.

[0] https://github.com/llvm/llvm-project/commit/ab391beb11f733b526b86f9df23734a34657d876
[1] https://github.com/aya-rs/aya/issues/1245