di: Remove the BTF map anonymization fixup
It was never necessary in the first place. There is no place in the
kernel that enforces BTF map structs themselves to be anonymous. Only
pointer types (that are members of BTF maps) have to be anonymous, but
that's handled by a separate fixup.

BTF maps work just fine without it.