ci: Install clang in version that matches LLVM version
LLVM bitcode is not backwards compatible and linking bitode objects
produced by different LLVM versions, even when it does not always
trigger any error in the backend, can lead to miscompilation (e.g.
https://github.com/aya-rs/bpf-linker/issues/318).

Using clang 15 for a compile test  was "working" out of luck, because
the program was simple enough to not cause a miscompilation. But given
that it's wrong, ensure that clang version matches LLVM version