ci: Specify target and static linking of C runtime explicitly
Without specifying the target, cargo uses the host target, which in case
of Linux runners is `*-unknown-linux-gnu`. To test static linking to the
full extent, use `*-unknown-linux-musl` targets.

On top of that, make sure that libc and C runtime are actually linked
statically by specifying `-C target-feature=+crt-static` rustflag. As of
today, that's still a default option on the most of `*-musl`
targets[0][1], but it's being phased out in favor of dynamic linking
being the default option.[2][3][4]

[0] https://github.com/rust-lang/rust/blob/672388edbee9e93c35e5fdf7dac818a6612a5103/compiler/rustc_target/src/spec/targets/x86_64_unknown_linux_musl.rs#L19-L20
[1] https://github.com/rust-lang/rust/blob/672388edbee9e93c35e5fdf7dac818a6612a5103/compiler/rustc_target/src/spec/targets/aarch64_unknown_linux_musl.rs#L17-L18
[2] https://github.com/rust-lang/compiler-team/issues/422
[3] https://github.com/rust-lang/rust/pull/133386
[4] https://github.com/rust-lang/rust/pull/14451